% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_var.R
\name{compute_var}
\alias{compute_var}
\title{Compute variance}
\usage{
compute_var(family, mu, phi, MC = FALSE)
}
\arguments{
\item{family}{character specifying the model. Currently are: \code{betaMode}, \code{unitGompertz},
\code{unitGamma} and \code{Kumaraswamy}}

\item{mu}{mode parameter}

\item{phi}{nuance parameter}

\item{MC}{logical indicating if use Monte Carlo simulations to compute variance.}
}
\description{
Compute analitycal or Monte Carlo variance for unit Modal distributions
}
\examples{
mu <- 0.5
phi <- 3.0

V_beta   <- compute_var("betaMode", mu, phi)
V_ugtz   <- compute_var("unitGompertz", mu, phi)
V_ugamma <- compute_var("unitGamma", mu, phi)
V_kum    <- compute_var("Kumaraswamy", mu, phi)

cbind(mu=mu, phi=phi, V_beta, V_ugtz, V_ugamma, V_kum)

V_beta   <- compute_var("betaMode", mu, phi, MC=TRUE)
V_ugtz   <- compute_var("unitGompertz", mu, phi, MC=TRUE)
V_ugamma <- compute_var("unitGamma", mu, phi, MC=TRUE)
V_kum    <- compute_var("Kumaraswamy", mu, phi, MC=TRUE)

cbind(mu=mu, phi=phi, V_beta, V_ugtz, V_ugamma, V_kum)

}
\author{
André Felipe B. Menezes \email{andrefelipemaringa@gmail.com}
}
