% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{Goodness-of-Fit}
\alias{Goodness-of-Fit}
\alias{gof}
\alias{voung.test}
\title{Goodness-of-Fit for models of class \code{unitModalReg}}
\usage{
gof(..., lt = NULL)

voung.test(object1, object2)
}
\arguments{
\item{...}{pass the models separate by commma to make the comparisons or use the argument \code{lt}.}

\item{lt}{list of models.}

\item{object1, object2}{obected of class \code{unitModalReg} contain the fitted models.}
}
\description{
Compute goodness-of-fit measures and Voung test for models of class \code{unitModalReg}.
}
\author{
André F. B. Menezes \email{andrefelipemaringa@gmail.com}
}
