% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hnp.R
\name{hnp}
\alias{hnp}
\alias{hnp.unitModalReg}
\title{Half-Normal Plot with Simulation Envelopes for \code{unitModalReg} Objects}
\usage{
hnp(object, ...)

\method{hnp}{unitModalReg}(
  object,
  nsim = 99,
  halfnormal = TRUE,
  plot = TRUE,
  level = 0.95,
  resid.type = c("cox-snell", "quantile"),
  ...
)
}
\arguments{
\item{object}{fitted model object of class \code{\link{unitModalReg}}.}

\item{...}{additional graphical parameters.}

\item{nsim}{number of simulations used to compute envelope. Default is 99.}

\item{halfnormal}{logical. If \code{TRUE}, a half-normal plot is produced. If \code{FALSE}, a normal plot is produced.}

\item{plot}{should the (half-)normal plot be plotted? Default is \code{TRUE}.}

\item{level}{confidence level of the simulated envelope. Default is 0.95.}

\item{resid.type}{type of residuals to be used: The currently options are \code{cox-snell} or \code{quantile}.}
}
\description{
Produces a (half-)normal plot from a fitted model object of class \code{\link{unitModalReg}}.
}
\author{
André F. B. Menezes \email{andrefelipemaringa@gmail.com}
}
