% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{methods-unitModalReg}
\alias{methods-unitModalReg}
\alias{print.unitModalReg}
\alias{summary.unitModalReg}
\alias{coef.unitModalReg}
\alias{vcov.unitModalReg}
\alias{logLik.unitModalReg}
\alias{confint.unitModalReg}
\alias{fitted.unitModalReg}
\title{Methods for \code{unitModalReg} Objects}
\usage{
\method{print}{unitModalReg}(x, digits = 4, ...)

\method{summary}{unitModalReg}(object, correlation = FALSE, ...)

\method{coef}{unitModalReg}(object, ...)

\method{vcov}{unitModalReg}(object, ...)

\method{logLik}{unitModalReg}(object, ...)

\method{confint}{unitModalReg}(object, parm, level = 0.95, ...)

\method{fitted}{unitModalReg}(object, ...)
}
\arguments{
\item{digits}{minimal number of \emph{significant} digits}

\item{...}{additional argument(s) for methods. Currently not used.}

\item{object, x}{fitted model object of class \code{\link{unitModalReg}}.}

\item{correlation}{logical; if \code{TRUE}, the correlation matrix of
the estimated parameters is returned and printed. Default is \code{FALSE}.}

\item{parm}{a specification of which parameters are to be given confidence intervals,
either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{the confidence level required.}
}
\description{
Methods for extracting information from fitted models objects of class \code{\link{unitModalReg}}.
}
\author{
André F. B. Menezes \email{andrefelipemaringa@gmail.com}
}
