% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.unitModalReg}
\alias{predict.unitModalReg}
\title{Prediction Method for \code{unitModalReg} Objects}
\usage{
\method{predict}{unitModalReg}(
  object,
  newdata = NULL,
  type,
  interval = FALSE,
  level = 0.95,
  se.fit = FALSE,
  ...
)
}
\arguments{
\item{object}{fitted model object of class \code{\link{unitModalReg}}.}

\item{newdata}{optionally, a data frame in which to look for variables with
which to predict. If omitted, the original observations are used.}

\item{type}{the type of prediction required. The default is on the scale of
the response variable. The "\code{terms}" option returns a matrix giving
the fitted values of each term in the model formula on the linear predictor
scale.}

\item{interval}{type of interval desired. The options are \code{none} and \code{confidence}.}

\item{level}{coverage probability for the confidence intervals. Default is \code{0.95}.}

\item{se.fit}{logical. If \code{TRUE} return the asymptotic standard errors.}

\item{...}{currently not used.}
}
\description{
Extract various types of predictions from fitted models objects
of class \code{\link{unitModalReg}}.
}
\author{
André F. B. Menezes \email{andrefelipemaringa@gmail.com}
}
