% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.unitModalReg}
\alias{residuals.unitModalReg}
\title{Residuals Method for \code{unitModalReg} Objects}
\usage{
\method{residuals}{unitModalReg}(
  object,
  type = c("cox-snell", "quantile", "pearson", "working", "response"),
  ...
)
}
\arguments{
\item{object}{fitted model object of class \code{\link{unitModalReg}}.}

\item{type}{character indicating type of residuals.}

\item{...}{currently not used.}
}
\description{
Extract various types of residuals from unit regression models: cox-snell and quantile residuals for now.
}
\author{
André F. B. Menezes \email{andrefelipemaringa@gmail.com}
}
