% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr.R
\name{ugamma}
\alias{ugamma}
\alias{dugamma}
\alias{pugamma}
\alias{qugamma}
\alias{rugamma}
\title{unit-Gamma distribution}
\usage{
dugamma(x, mu, phi, log = FALSE)

pugamma(q, mu, phi, lower.tail = TRUE, log.p = FALSE)

qugamma(p, mu, phi, lower.tail = TRUE, log.p = FALSE)

rugamma(n, mu, phi)
}
\arguments{
\item{x, q}{vector of positive quantiles.}

\item{mu}{location parameter indicating the mode.}

\item{phi}{shape parameter.}

\item{log, log.p}{logical; If TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; If TRUE, (default), \eqn{P(X \leq x)} are returned, otherwise \eqn{P(X > x)}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}
}
\value{
\code{dugamma} gives the density, \code{pugamma} gives the distribution function,
\code{qugamma} gives the quantile function and \code{rugamma} generates random deviates.

Invalid arguments will return an error message.
}
\description{
Density function, distribution function, quantile function, random number generation function
for the unit-Gamma distribution re-parametrized in terms of the mode.
}
\examples{
set.seed(6969)
x <- rugamma(n = 1e4, mu = 0.5, phi = 3.0)
R <- range(x)
S <- seq(from = R[1], to = R[2], by = 0.01)
hist(x, prob = TRUE, main = 'unit-gamma distribution')
lines(S, dugamma(x = S, mu = 0.5, phi = 3.0), col = 2)
plot(ecdf(x))
lines(S, pugamma(q = S, mu = 0.5, phi = 3.0), col = 2)
plot(quantile(x, probs = S), type = "l")
lines(qugamma(p = S, mu = 0.5, phi = 3.0), col = 2)
}
\author{
André Felipe B. Menezes \email{andrefelipemaringa@gmail.com}
}
