% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr.R
\name{umaxwell}
\alias{umaxwell}
\alias{dumaxwell}
\alias{pumaxwell}
\alias{qumaxwell}
\alias{rumaxwell}
\title{unit-maxwell distribution}
\usage{
dumaxwell(x, mu, log = FALSE)

pumaxwell(q, mu, lower.tail = TRUE, log.p = FALSE)

qumaxwell(p, mu, lower.tail = TRUE, log.p = FALSE)

rumaxwell(n, mu)
}
\arguments{
\item{x, q}{vector of positive quantiles.}

\item{mu}{location parameter indicating the mode.}

\item{log, log.p}{logical; If TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; If TRUE, (default), \eqn{P(X \leq x)} are returned, otherwise \eqn{P(X > x)}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}
}
\value{
\code{dumaxwell} gives the density, \code{pumaxwell} gives the distribution function,
\code{qumaxwell} gives the quantile function and \code{rumaxwell} generates random deviates.

Invalid arguments will return an error message.
}
\description{
Density function, distribution function, quantile function, random number generation function
for the unit-maxwell distribution re-parametrized in terms of the mode.
}
\examples{
set.seed(6969)
x <- rumaxwell(n = 1e4, mu = 0.5)
R <- range(x)
S <- seq(from = R[1], to = R[2], by = 0.01)
hist(x, prob = TRUE, main = 'unit-maxwell distribution')
lines(S, dumaxwell(x = S, mu = 0.5), col = 2)
plot(ecdf(x))
lines(S, pumaxwell(q = S, mu = 0.5), col = 2)
plot(quantile(x, probs = S), type = "l")
lines(qumaxwell(p = S, mu = 0.5), col = 2)
}
\author{
André Felipe B. Menezes \email{andrefelipemaringa@gmail.com}
}
