% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitModalReg.R
\name{unitModalReg.control}
\alias{unitModalReg.control}
\title{Control Parameters for unit modal regression models}
\usage{
unitModalReg.control(method = "BFGS", maxit = 5000, reltol = 1e-08, ...)
}
\arguments{
\item{method}{characters string specifying the method argument passed to \code{optim}.}

\item{maxit}{integer specifying the \code{maxit} argument (maximal number of iterations) passed to \code{optim}.}

\item{reltol}{relative convergence tolerance passed to \code{optim}.}

\item{...}{arguments passed to \code{optim}.}
}
\description{
Parameters that control fitting of unit modal regression models using in \code{unitModalReg}.
}
\author{
André F. B. Menezes \email{andrefelipemaringa@gmail.com}
}
