% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitModalReg.R
\name{unitModalReg}
\alias{unitModalReg}
\alias{unitModalReg.fit}
\title{Unit Modal Regression Models}
\usage{
unitModalReg(
  formula,
  data,
  family,
  link,
  start = NULL,
  control = unitModalReg.control()
)

unitModalReg.fit(y, X, family, link, start, control = unitModalReg.control())
}
\arguments{
\item{formula}{symbolic description of the quantile model.}

\item{data}{data.frame contain the variables in the model.}

\item{family}{specify the distribution family name.}

\item{link}{character specification of the link function in the modal regression. Default is \code{logit}.}

\item{start}{an optional vector with starting values for all parameters}

\item{control}{a list of control arguments specified via \code{unitModalReg.control}.}

\item{y}{numeric vector of response variable.}

\item{X}{numeric matrix. Design matrix for the modal regression.}
}
\description{
Fit several unit modal regression models.
}
\author{
André F. B. Menezes \email{andrefelipemaringa@gmail.com}
}
